+++ packages/ai-editor/src/browser/index.ts
@@ -16,5 +16,3 @@
 
 export * from './ai-editor-command-contribution';
 export * from './ask-ai-input-widget';
-export * from './editor-context-collector';
-export * from './editor-context-collector-service';
